Rename nested signals in library Fixes #478
