+++ sdk/go/consts.go
@@ -8,7 +8,7 @@ const (
     DatastarKey = "datastar"
     Version                   = "1.0.0-beta.1"
     VersionClientByteSize     = 36235
-    VersionClientByteSizeGzip = 13244
+    VersionClientByteSizeGzip = 13242
 
     //region Default durations
